# src/app/api/openapi.yaml
openapi: 3.0.0
info:
  title: ZenZen CRM API
  version: '1.0.0'
  description: |
    API for the ZenZen CRM system, providing ticket management, messaging, and team collaboration features.
  contact:
    name: ZenZen Support
    url: https://zenzen.support
servers:
  - url: https://api.zenzen.dev
    description: Development server
  - url: https://api.zenzen.com
    description: Production server
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Ticket:
      type: object
      properties:
        id:
          type: string
          format: uuid
        customer_id:
          type: string
          format: uuid
        title:
          type: string
        status:
          type: string
          enum: [UNOPENED, IN PROGRESS, RESOLVED, UNRESOLVED]
        priority:
          type: string
          enum: [NONE, LOW, MEDIUM, HIGH, CRITICAL]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        first_response_at:
          type: string
          format: date-time
        resolved_at:
          type: string
          format: date-time
        reopen_count:
          type: integer
        assigned_to:
          type: string
          format: uuid
        assigned_team:
          type: string
          format: uuid
        tags:
          type: array
          items:
            type: string
        custom_fields:
          type: object
          additionalProperties: true
...
    Template:
      type: object
      properties:
        id:
          type: string
          format: uuid
        created_by:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
        created_at:
          type: string
          format: date-time
...
paths:
...
  /tickets:
    get:
      summary: List tickets
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [UNOPENED, IN PROGRESS, RESOLVED, UNRESOLVED]
        - name: priority
          in: query
          schema:
            type: string
            enum: [NONE, LOW, MEDIUM, HIGH, CRITICAL]
        - name: assigned_to
          in: query
          schema:
            type: string
            format: uuid
        - name: assigned_team
          in: query
          schema:
            type: string
            format: uuid
        - name: customer_id
          in: query
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of tickets
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Ticket'
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                      pages:
                        type: integer
                      current_page:
                        type: integer
                      per_page:
                        type: integer

    post:
      summary: Create a new ticket
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                priority:
                  type: string
                  enum: [NONE, LOW, MEDIUM, HIGH, CRITICAL]
                tags:
                  type: array
                  items:
                    type: string
                custom_fields:
                  type: object
                  additionalProperties: true
      responses:
        '201':
          description: Ticket created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
...
  /templates:
    get:
      summary: List response templates
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: created_by
          in: query
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of templates
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Template'
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                      pages:
                        type: integer
                      current_page:
                        type: integer
                      per_page:
                        type: integer

    post:
      summary: Create a new template
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - content
              properties:
                title:
                  type: string
                content:
                  type: string
      responses:
        '201':
          description: Template created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
...
security:
  - ApiKeyAuth: []
  - BearerAuth: [] 